요청사항 변경

+++ api/src/main/java/com/ugly/agri/domain/Product.java
@@ -4,6 +4,7 @@ import com.ugly.agri.type.CategoryType;
 import lombok.*;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name = "products")
@@ -21,7 +22,7 @@ public class Product extends BaseEntity {
     private String name;
 
     @Column(name = "price", nullable = false)
-    private Long price;
+    private Integer price;
 
     @Column(name = "comment", nullable = false)
     private String comment;
